Add starknet-artifacts crate
