+++ crates/cheatnet/Cargo.toml
@@ -30,6 +30,7 @@ num-bigint.workspace = true
 conversions.workspace = true
 fs2 = "0.4.3"
 flate2 = "1.0.27"
+starknet-artifacts = { path = "../starknet-artifacts" }
 
 [dev-dependencies]
 ctor.workspace = true
