+++ crates/forge/src/lib.rs
@@ -28,8 +28,10 @@ use once_cell::sync::Lazy;
 use rand::{thread_rng, RngCore};
 use smol_str::SmolStr;
 
+use starknet_artifacts::StarknetContractArtifacts;
+
 use crate::fuzzer::RandomFuzzer;
-use crate::scarb::{ForgeConfig, ForkTarget, StarknetContractArtifacts};
+use crate::scarb::{ForgeConfig, ForkTarget};
 
 // pub use crate::collecting::CrateLocation;
 pub use crate::test_crate_summary::TestCrateSummary;
