+++ crates/forge/tests/integration/common/corelib.rs
@@ -1,9 +1,9 @@
 use assert_fs::fixture::{FileTouch, FileWriteStr, PathChild, PathCopy};
 use assert_fs::TempDir;
 use camino::Utf8PathBuf;
-use forge::scarb::corelib_for_package;
 use indoc::indoc;
 use scarb_metadata::MetadataCommand;
+use starknet_artifacts::corelib_for_package;
 
 #[allow(clippy::module_name_repetitions)]
 pub fn corelib_path() -> Utf8PathBuf {
