+++ crates/forge/tests/integration/common/runner.rs
@@ -11,8 +11,8 @@ use cairo_lang_starknet::contract_class::compile_contract_in_prepared_db;
 use cairo_lang_starknet::inline_macros::selector::SelectorMacro;
 use cairo_lang_starknet::plugin::StarkNetPlugin;
 use camino::{Utf8Path, Utf8PathBuf};
-use forge::scarb::StarknetContractArtifacts;
 use forge::{CrateLocation, TestCrateSummary};
+use starknet_artifacts::StarknetContractArtifacts;
 use std::collections::HashMap;
 use std::fs;
 use std::path::{Path, PathBuf};
