+++ crates/starknet-artifacts/tests/data/panic_decoding/Scarb.toml
@@ -1,0 +1,3 @@
+[package]
+name = "panic_decoding"
+version = "0.1.0"
