+++ crates/starknet-artifacts/tests/data/simple_package/Scarb.toml
@@ -1,0 +1,14 @@
+[package]
+name = "simple_package"
+version = "0.1.0"
+
+[dependencies]
+starknet = "2.2.0"
+snforge_std = { path = "../../../../../snforge_std" }
+
+[[target.starknet-contract]]
+sierra = true
+casm = true
+
+[tool.snforge]
+exit_first = false
