+++ crates/starknet-artifacts/tests/data/simple_package/tests/without_prefix.cairo
@@ -1,0 +1,4 @@
+#[test]
+fn five() {
+    assert(5 == 5, '5 == 5');
+}
